moved HTTPNetworkingServerError to its own file + tests
